Ready 4 homework

+++ ru.tinkoff/java/bot/src/main/java/ru/tinkoff/edu/java/bot/telegram/commands/UntrackCommand.java
@@ -13,6 +13,7 @@ public class UntrackCommand implements Command {
     private String DESCRIPTION = COMMAND + " -> перестать отслеживать ссылку.";
 
     private String GOOD_ANSWER = "Ссылка успешна удалена";
+    private String GOOD_ANSWER_BEFORE = "Введите ссылку в формате URL, которую хотите прекратить отслеживать";
     private String BAD_ANSWER = "Такой ссылки не существует, бот поддерживает ссылки в формате URL";
 
     public UntrackCommand(LinksClient linksClient) {
@@ -32,6 +33,8 @@ public class UntrackCommand implements Command {
 
     @Override
     public SendMessage handle(Update update) {
+        var msg = update.message().text();
+        if (msg.toCharArray()[0] == '/') return new SendMessage(update.message().chat().id(), GOOD_ANSWER_BEFORE);
         linksClient.deleteLink(update.message().chat().id(), new RemoveLinkRequest(update.message().text()));
         return new SendMessage(update.message().chat().id(), GOOD_ANSWER);
     }
